feat: update promotion retrieval logic to include user role checks and remove unused endpoints
